Moved  config to resources and applied it in movement and shooting.
